requirements.txt
python-dotenv==0.21.1
requests==2.31.0
lxml==5.2.2
psycopg2==2.7.7 ; sys_platform != 'win32' and python_version < '3.8'
psycopg2==2.8.6 ; sys_platform != 'win32' and python_version >= '3.8'and python_version <= '3.10'
psycopg2==2.8.6 ; sys_platform == 'win32' and python_version < '3.10'
psycopg2==2.9.5 ; python_version > '3.10' or ( sys_platform == 'win32' and python_version == '3.10')
prettytable==3.10.2